requirement.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
numpy==1.26.4
pygame==2.5.2
pyserial==3.5
requests==2.32.2
scipy==1.13.1
urllib3==2.2.1
